requirements.txt
appnope==0.1.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.2
fonttools==4.29.0
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.22.1
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.5.0
plotly==5.6.0
plotly-express==0.4.1
prompt-toolkit==3.0.27
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
scipy==1.8.0
six==1.16.0
stack-data==0.1.4
statsmodels==0.13.2
tenacity==8.0.1
tomli==2.0.1
torch==1.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
wcwidth==0.2.5
